# server/openapi.yaml
openapi: 3.0.3
info:
  title: ShieldXChurn API
  description: >
    Churn prediction, risk analysis, and retention campaign management
    for Aspora's cross-border remittance platform.
  version: 1.0.0
  contact:
    name: Aspora Engineering
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/interventions/trigger:
    post:
      tags: [Campaigns]
      summary: Trigger intervention for a user
      operationId: triggerIntervention
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                channel:
                  type: string
                campaign:
                  type: string
                playbook:
                  type: string
      responses:
        '200':
          description: Intervention triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionLogEntry'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/interventions/outcome:
    post:
      tags: [Campaigns]
      summary: Record intervention outcome
      operationId: recordOutcome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intervention_id, outcome]
              properties:
                intervention_id:
                  type: string
                outcome:
                  type: string
                  enum: [retained, churned, no_response]
                notes:
                  type: string
      responses:
        '200':
          description: Updated intervention entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionLogEntry'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/simulator:
    post:
      tags: [Campaigns]
      summary: Revenue impact simulator
      operationId: runSimulator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cohort_key:
                  type: string
                playbook:
                  type: string
                budget:
                  type: number
                target_count:
                  type: integer
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulatorResult'
...
  /api/moengage/engage:
    post:
      tags: [Campaigns]
      summary: Send single intervention via MoEngage
      operationId: moengageEngage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                channel:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Engagement result
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: MoEngage send failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/moengage/bulk:
    post:
      tags: [Campaigns]
      summary: Bulk campaign via MoEngage
      operationId: moengageBulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tier:
                  type: string
                  enum: [CRITICAL, HIGH, MEDIUM, LOW]
                max_count:
                  type: integer
                channel:
                  type: string
      responses:
        '200':
          description: Bulk campaign result
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Bulk campaign failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/retell/call:
    post:
      tags: [Campaigns]
      summary: Initiate AI voice call via Retell
      operationId: retellCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: Call initiated
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Call failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/metabase/query:
    post:
      tags: [Campaigns]
      summary: Redshift query proxy (SELECT only)
      operationId: metabaseQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sql:
                  type: string
                  description: Raw SELECT query
                query_key:
                  type: string
                  description: Named query shortcut
      responses:
        '200':
          description: Query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items:
                      type: string
                  rows:
                    type: array
                    items:
                      type: object
                  row_count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          description: Query failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- AI ---
...
components:
...
  schemas:
...
    SimulatorResult:
      type: object
      properties:
        cohort:
          type: string
        playbook:
          type: string
        target_users:
          type: integer
        avg_churn_probability:
          type: number
        lift_rate:
          type: number
        expected_retained:
          type: integer
        projected_revenue:
          type: integer
        projected_cost:
          type: integer
        net_profit:
          type: integer
        roi:
          type: number
    InterventionLogEntry:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        channel:
          type: string
        campaign:
          type: string
        playbook:
          type: string
        risk_tier:
          type: string
        churn_probability:
          type: number
        triggered_at:
          type: string
        status:
          type: string
        outcome:
          type: string
          nullable: true
        outcome_at:
          type: string
        notes:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'